Tests for YouTubeVideo.
